add getall&getrole

+++ controllers/userController.js
@@ -14,7 +14,7 @@ exports.createUser = async (req, res) => {
     });
 
     if (existingUser) {
-      return res.status(400).send('มีผู้ใช้ที่มีชื่อผู้ใช้นี้อยู่แล้ว');
+      return res.status(400).send('There is a user who already has this username.');
     }
 
     // Hash the password
@@ -60,14 +60,40 @@ exports.getAllUsers = async (req, res) => {
   }
 };
 
+// Get Roles 
+exports.getRole = async (req, res) => {
+  const { role } = req.params;
+
+  try {
+      const users = await prisma.user.findMany({
+          where: {
+              role: role.toUpperCase() // ตรวจสอบให้แน่ใจว่าบทบาทถูกแปลงเป็นตัวพิมพ์ใหญ่เพื่อให้ตรงกับฐานข้อมูล
+          }
+      });
+
+      res.status(200).json(users);
+  } catch (error) {
+      console.error("Error fetching users by role:", error);
+      res.status(400).json({ error: "Error fetching users by role:" });
+  }
+};
 
 
-// Update User
+// Update User   Error fetching users by role:
 exports.updateUser = async (req, res) => {
   const { id } = req.params;
   const { name, username, password } = req.body;
 
   try {
+    // Check if the user exists
+    const existingUser = await prisma.user.findUnique({
+      where: { id: parseInt(id) },
+    });
+
+    if (!existingUser) {
+      return res.status(404).json({ message: `User with ID ${id} not found` });
+    }
+
     const hashedPassword = password ? await bcrypt.hash(password, 10) : undefined;
 
     const updatedUser = await prisma.user.update({
@@ -79,18 +105,29 @@ exports.updateUser = async (req, res) => {
       },
     });
 
-    res.status(200).json({ message: `User with ID ${id} has Update successfully` });
+    res.status(200).json({ message: `User with ID ${id} has been updated successfully` });
   } catch (error) {
     console.error("Error updating user:", error.message);
     res.status(400).json({ error: error.message });
   }
 };
 
+
+
 // Delete User
 exports.deleteUser = async (req, res) => {
   const { id } = req.params;
 
   try {
+    // Check if the user exists before deleting
+    const existingUser = await prisma.user.findUnique({
+      where: { id: parseInt(id) },
+    });
+
+    if (!existingUser) {
+      return res.status(404).json({ message: `User with ID ${id} not found` });
+    }
+
     await prisma.user.delete({
       where: { id: parseInt(id) },
     });
@@ -101,7 +138,6 @@ exports.deleteUser = async (req, res) => {
     res.status(400).json({ error: error.message });
   }
 };
-
